Fix possible int overflow in io.fits compression

+++ astropy/io/fits/hdu/compressed/src/compression.c
@@ -320,7 +320,7 @@ static PyObject *compress_hcompress_1_c(PyObject *self, PyObject *args) {
     return (PyObject *)NULL;
   }
 
-  if (count != nx * ny * bytepix) {
+  if (count != (Py_ssize_t) nx * ny * bytepix) {
     PyErr_SetString(PyExc_ValueError,
                     "The tile dimensions and dtype do not match the number of bytes provided.");
     return (PyObject *)NULL;
@@ -395,7 +395,7 @@ static PyObject *decompress_hcompress_1_c(PyObject *self, PyObject *args) {
 
   compressed_values = (unsigned char *)str;
 
-  dbytes = malloc(nx * ny * bytepix);
+  dbytes = malloc((size_t) nx * ny * bytepix);
 
   if (bytepix == 4) {
     decompressed_values_int = (int *)dbytes;
@@ -421,7 +421,7 @@ static PyObject *decompress_hcompress_1_c(PyObject *self, PyObject *args) {
   }
 
   // fits_hdecompress[64] always returns 4 byte integers
-  result = Py_BuildValue("y#", dbytes, nx * ny * 4);
+  result = Py_BuildValue("y#", dbytes, (Py_ssize_t) nx * ny * 4);
   free(dbytes);
   return result;
 }
